+++ packages/video_player_avplay/tizen/src/messages.cc
@@ -2041,6 +2041,48 @@ void VideoPlayerAvplayApi::SetUp(flutter::BinaryMessenger* binary_messenger,
   {
     BasicMessageChannel<> channel(binary_messenger,
                                   "dev.flutter.pigeon.video_player_avplay."
+                                  "VideoPlayerAvplayApi.updateDashToken" +
+                                      prepended_suffix,
+                                  &GetCodec());
+    if (api != nullptr) {
+      channel.SetMessageHandler(
+          [api](const EncodableValue& message,
+                const flutter::MessageReply<EncodableValue>& reply) {
+            try {
+              const auto& args = std::get<EncodableList>(message);
+              const auto& encodable_player_id_arg = args.at(0);
+              if (encodable_player_id_arg.IsNull()) {
+                reply(WrapError("player_id_arg unexpectedly null."));
+                return;
+              }
+              const int64_t player_id_arg = encodable_player_id_arg.LongValue();
+              const auto& encodable_dash_token_arg = args.at(1);
+              if (encodable_dash_token_arg.IsNull()) {
+                reply(WrapError("dash_token_arg unexpectedly null."));
+                return;
+              }
+              const auto& dash_token_arg =
+                  std::get<std::string>(encodable_dash_token_arg);
+              ErrorOr<bool> output =
+                  api->UpdateDashToken(player_id_arg, dash_token_arg);
+              if (output.has_error()) {
+                reply(WrapError(output.error()));
+                return;
+              }
+              EncodableList wrapped;
+              wrapped.push_back(EncodableValue(std::move(output).TakeValue()));
+              reply(EncodableValue(std::move(wrapped)));
+            } catch (const std::exception& exception) {
+              reply(WrapError(exception.what()));
+            }
+          });
+    } else {
+      channel.SetMessageHandler(nullptr);
+    }
+  }
+  {
+    BasicMessageChannel<> channel(binary_messenger,
+                                  "dev.flutter.pigeon.video_player_avplay."
                                   "VideoPlayerAvplayApi.getActiveTrackInfo" +
                                       prepended_suffix,
                                   &GetCodec());